# frontend/src/data/openapi-spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Store CRUD API",
    "version": "1.0.0",
    "description": "A comprehensive REST API for managing store inventory, dynamic tables, and e-commerce functionality.",
    "contact": {
      "name": "API Support"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8787",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/api/items": {
      "get": {
        "tags": ["Items"],
        "summary": "List Store Items",
        "description": "Retrieve a paginated list of items with filtering and search capabilities",
        "security": [{"BearerAuth": []}],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "Page number for pagination",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of items",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Item"}
                    },
                    "total": {"type": "integer"},
                    "page": {"type": "integer"},
                    "totalPages": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Items"],
        "summary": "Create New Item",
        "description": "Add a new item to the inventory",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CreateItemRequest"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Item created successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Item"}
              }
            }
          }
        }
      }
    },
...
    "/api/tables": {
      "get": {
        "tags": ["Tables"],
        "summary": "List Tables",
        "description": "Retrieve user's dynamic tables with metadata",
        "security": [{"BearerAuth": []}],
        "responses": {
          "200": {
            "description": "List of tables",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {"$ref": "#/components/schemas/Table"}
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Tables"],
        "summary": "Create Table",
        "description": "Create a new dynamic table with custom schema",
        "security": [{"BearerAuth": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"$ref": "#/components/schemas/CreateTableRequest"}
            }
          }
        },
        "responses": {
          "201": {
            "description": "Table created successfully",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Table"}
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Bearer token authentication. Format: `Bearer <token>`"
      }
    },
    "schemas": {
      "Item": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique item identifier"
          },
          "name": {
            "type": "string",
            "description": "Item name"
          },
          "description": {
            "type": "string",
            "description": "Item description",
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "float",
            "description": "Item price"
          },
          "image_url": {
            "type": "string",
            "format": "uri",
            "description": "URL to item image",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateItemRequest": {
        "type": "object",
        "required": ["name", "price"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Item name"
          },
          "description": {
            "type": "string",
            "description": "Item description"
          },
          "price": {
            "type": "number",
            "format": "float",
            "minimum": 0,
            "description": "Item price"
          },
          "image_url": {
            "type": "string",
            "format": "uri",
            "description": "URL to item image"
          }
        }
      },
...
      "Table": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Unique table identifier"
          },
          "name": {
            "type": "string",
            "description": "Table name"
          },
          "description": {
            "type": "string",
            "description": "Table description"
          },
          "for_sale": {
            "type": "boolean",
            "description": "Whether this is an e-commerce table"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateTableRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "description": "Table name"
          },
          "description": {
            "type": "string",
            "description": "Table description"
          },
          "for_sale": {
            "type": "boolean",
            "description": "Whether this is an e-commerce table",
            "default": false
          }
        }
      }
    }
  },
  "security": [
    {
      "BearerAuth": []
    }
  ]
}